+++ src/lib/flows/full-downstream-flow.js
@@ -25,7 +25,13 @@ async function start(context, options = { isArchiveArtifacts: true }) {
   const nodeChain = await getOrderedListForProject(
     context.config.github.inputs.definitionFile,
     projectTriggeringJob,
-    await getPlaceHolders(context, context.config.github.inputs.definitionFile)
+    {
+      urlPlaceHolders: await getPlaceHolders(
+        context,
+        context.config.github.inputs.definitionFile
+      ),
+      token: context.token
+    }
   );
 
   logger.info(