+++ src/lib/flows/full-downstream-flow.js
@@ -14,7 +14,19 @@ const {
   archiveArtifacts
 } = require("../artifacts/build-chain-flow-archive-artifact-helper");
 
+const { execute: executePre } = require("./sections/pre");
+const { execute: executePost } = require("./sections/post");
+
 async function start(context, options = { isArchiveArtifacts: true }) {
+  const readerOptions = {
+    urlPlaceHolders: await getPlaceHolders(
+      context,
+      context.config.github.inputs.definitionFile
+    ),
+    token: context.token
+  };
+  await executePre(context.config.github.inputs.definitionFile, readerOptions);
+
   core.startGroup(
     `[Full Downstream Flow] Checking out ${context.config.github.groupProject} and its dependencies`
   );
@@ -25,13 +37,7 @@ async function start(context, options = { isArchiveArtifacts: true }) {
   const nodeChain = await getOrderedListForProject(
     context.config.github.inputs.definitionFile,
     projectTriggeringJob,
-    {
-      urlPlaceHolders: await getPlaceHolders(
-        context,
-        context.config.github.inputs.definitionFile
-      ),
-      token: context.token
-    }
+    readerOptions
   );
 
   logger.info(
@@ -71,6 +77,12 @@ async function start(context, options = { isArchiveArtifacts: true }) {
   } else {
     logger.info("Archive artifact won't be executed");
   }
+
+  await executePost(
+    context.config.github.inputs.definitionFile,
+    executionResult,
+    readerOptions
+  );
 
   if (executionResult !== true) {
     logger.error(executionResult);