+++ src/lib/flows/pull-request-flow.js
@@ -13,11 +13,22 @@ const core = require("@actions/core");
 const {
   archiveArtifacts
 } = require("../artifacts/build-chain-flow-archive-artifact-helper");
+const { execute: executePre } = require("./sections/pre");
+const { execute: executePost } = require("./sections/post");
 
 async function start(
   context,
   options = { skipProjectCheckout: new Map(), isArchiveArtifacts: true }
 ) {
+  const readerOptions = {
+    urlPlaceHolders: await getPlaceHolders(
+      context,
+      context.config.github.inputs.definitionFile
+    ),
+    token: context.token
+  };
+  await executePre(context.config.github.inputs.definitionFile, readerOptions);
+
   core.startGroup(
     `[Pull Request Flow] Checking out ${context.config.github.groupProject} and its dependencies`
   );
@@ -28,13 +39,7 @@ async function start(
   const definitionTree = await getTreeForProject(
     context.config.github.inputs.definitionFile,
     projectTriggeringJob,
-    {
-      urlPlaceHolders: await getPlaceHolders(
-        context,
-        context.config.github.inputs.definitionFile
-      ),
-      token: context.token
-    }
+    readerOptions
   );
   const nodeChain = await parentChainFromNode(definitionTree);
   logger.info(
@@ -74,6 +79,12 @@ async function start(
   } else {
     logger.info("Archive artifact won't be executed");
   }
+
+  await executePost(
+    context.config.github.inputs.definitionFile,
+    executionResult,
+    readerOptions
+  );
 
   if (executionResult !== true) {
     logger.error(executionResult);