+++ test/lib/flows/pull-request-flow.test.js
@@ -26,6 +26,13 @@ const {
 } = require("@kie/build-chain-configuration-reader");
 jest.mock("@kie/build-chain-configuration-reader");
 
+const { execute: executePre } = require("../../../src/lib/flows/sections/pre");
+const {
+  execute: executePost
+} = require("../../../src/lib/flows/sections/post");
+jest.mock("../../../src/lib/flows/sections/pre");
+jest.mock("../../../src/lib/flows/sections/post");
+
 afterEach(() => {
   jest.clearAllMocks();
 });
@@ -110,6 +117,23 @@ test("start no parent dependencies. project triggering the job", async () => {
     [{ project }, { project: "project2" }],
     ["success", "always"]
   );
+
+  expect(executePre).toHaveBeenCalledTimes(1);
+  expect(executePre).toHaveBeenCalledWith(
+    "test/resources/build-config/build-config.yaml",
+    {
+      token: undefined,
+      urlPlaceHolders: {}
+    }
+  );
+  expect(executePost).toHaveBeenCalledTimes(1);
+  expect(
+    executePost
+  ).toHaveBeenCalledWith(
+    "test/resources/build-config/build-config.yaml",
+    true,
+    { token: undefined, urlPlaceHolders: {} }
+  );
 });
 
 test("start no parent dependencies. project triggering the job. isArchiveArtifacts to false", async () => {
@@ -187,6 +211,23 @@ test("start no parent dependencies. project triggering the job. isArchiveArtifac
     { isArchiveArtifacts: false }
   );
   expect(archiveArtifacts).toHaveBeenCalledTimes(0);
+
+  expect(executePre).toHaveBeenCalledTimes(1);
+  expect(executePre).toHaveBeenCalledWith(
+    "test/resources/build-config/build-config.yaml",
+    {
+      token: undefined,
+      urlPlaceHolders: {}
+    }
+  );
+  expect(executePost).toHaveBeenCalledTimes(1);
+  expect(
+    executePost
+  ).toHaveBeenCalledWith(
+    "test/resources/build-config/build-config.yaml",
+    true,
+    { token: undefined, urlPlaceHolders: {} }
+  );
 });
 
 test("start no parent dependencies. project triggering the job. Execute Exception", async () => {
@@ -276,6 +317,23 @@ test("start no parent dependencies. project triggering the job. Execute Exceptio
     [{ project }, { project: "project2" }],
     ["failure", "always"]
   );
+
+  expect(executePre).toHaveBeenCalledTimes(1);
+  expect(executePre).toHaveBeenCalledWith(
+    "test/resources/build-config/build-config.yaml",
+    {
+      token: undefined,
+      urlPlaceHolders: {}
+    }
+  );
+  expect(executePost).toHaveBeenCalledTimes(1);
+  expect(
+    executePost
+  ).toHaveBeenCalledWith(
+    "test/resources/build-config/build-config.yaml",
+    new Error("error executing command"),
+    { token: undefined, urlPlaceHolders: {} }
+  );
 });
 
 test("start no parent dependencies. startingProject", async () => {
@@ -358,5 +416,22 @@ test("start no parent dependencies. startingProject", async () => {
     { project: "project2" },
     [{ project }, { project: "project2" }],
     ["success", "always"]
+  );
+
+  expect(executePre).toHaveBeenCalledTimes(1);
+  expect(executePre).toHaveBeenCalledWith(
+    "test/resources/build-config/build-config.yaml",
+    {
+      token: undefined,
+      urlPlaceHolders: {}
+    }
+  );
+  expect(executePost).toHaveBeenCalledTimes(1);
+  expect(
+    executePost
+  ).toHaveBeenCalledWith(
+    "test/resources/build-config/build-config.yaml",
+    true,
+    { token: undefined, urlPlaceHolders: {} }
   );
 });