+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `custom-command-treatment` CLI option
 - `@kie/build-chain-configuration-reader` upgraded to `2.0.4`
+- error mapping branches solved
 
 # V2.4