+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `@kie/build-chain-configuration-reader` upgraded to `2.0.4`
 - error mapping branches solved
 - allow to read definition files from private repositories
+- pre/post sections
 
 # V2.4